+++ func/pvOnline.php
@@ -6,6 +6,8 @@ if (!defined('pvault_panel')){ die('Not Found');}
 function pvPost($url, $data){
     $curl = curl_init($url);
     curl_setopt($curl, CURLOPT_POST, true);
+	curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 0);
+	curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
     curl_setopt($curl, CURLOPT_POSTFIELDS, http_build_query($data));
     curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
     return $response = curl_exec($curl);