+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ### Version 8.9
+- Fix a bug failing to populate soem fields when adding a supplier
 - Update PVOnline api v2
 - Fix api log file error
 - Add option to wipe out all ingredients